+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2018-06-16
 ### Added
 - `Scratchpad::static_new_in_place()` for initializing a `Scratchpad` within
   a block of uninitialized memory for `Scratchpad` types that use only static
@@ -175,7 +175,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/okready/scratchpad/compare/v1.0.1...HEAD
+[1.1.0]: https://github.com/okready/scratchpad/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/okready/scratchpad/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.2...v1.0.0
 [1.0.0-beta.2]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.1...v1.0.0-beta.2